Replace "replace" method with function. Smaller, tider, no reuse, slower
Normal single-replace should be fine, but the logic for continuous replacement is a little complicated. May be slightly slower, may have odd bugs. However, it's a lot tidier and easier to figure out what's going on than the old version was (180 lines > 20!)
This has the additional benefit of supporting all keypresses that send a character, rather than just the specific mapped keys.